+++ application/views/history.php
@@ -1,0 +1,16 @@
+
+<!-- Author: Aing Ragunathan
+
+This file holds the formatting for the view output. -->
+
+
+<div class="row">
+    {history}
+    	<div class="col-xs-1"><p>{date}</p></div>
+    	<div class="col-xs-2"><p>{type}</p></div>
+        <div class="col-xs-3"><p>{data}</p></div>
+        <div class="col-xs-5"><p>{location}</p></div>
+        <div class="col-xs-6"><p>Cost: {cost}</p></div>
+        <div class="col-xs-7"><p>Price: {price}</p></div>
+    {/history}
+</div>